requirements.txt
absl-py==2.0.0
aioredis==1.3.1
amqp==5.1.1
asgiref==3.5.2
astroid==2.15.0
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
autobahn==21.3.1
Automat==20.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
billiard==3.6.4.0
blinker==1.6.3
cachetools==5.3.1
celery==5.2.7
certifi==2022.12.7
cffi==1.14.6
channels==3.0.4
channels-redis==3.3.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cmake==3.21.3
constantly==15.1.0
cryptography==3.4.8
cycler==0.10.0
daphne==3.0.2
deepface==0.0.79
dill==0.3.6
Django==2.2.9
django-celery-beat==2.4.0
django-celery-results==2.4.0
django-detect==1.0.20
django-extensions==3.2.0
django-mobile-detect==0.1
django-rest-authtoken==2.1.4
django-six==1.0.4
django-timezone-field==5.0
djangorestframework==3.13.1
dlib==19.24.99
elastic-transport==8.4.0
elasticsearch==8.6.1
et-xmlfile==1.1.0
exceptiongroup==1.1.0
face-recognition==1.3.0
face-recognition-models==0.3.0
filelock==3.9.0
fire==0.5.0
flask==3.0.0
flatbuffers==23.5.26
flower==1.2.0
gast==0.4.0
gdown==4.7.1
google-auth==2.23.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.59.0
gunicorn==21.2.0
h11==0.14.0
h5py==3.10.0
hiredis==2.0.0
huggingface-hub==0.12.1
humanize==4.6.0
hyperlink==21.0.0
idna==3.4
importlib-metadata==6.8.0
imutils==0.5.2
incremental==21.3.0
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
JPype1==1.4.1
JSON-log-formatter==0.5.1
keras==2.13.1
kiwisolver==1.3.2
kombu==5.2.4
konlpy==0.6.0
lazy-object-proxy==1.9.0
libclang==16.0.6
lxml==4.9.2
Markdown==3.5
MarkupSafe==2.1.3
matplotlib==3.4.3
mccabe==0.7.0
mecab-python===0.996-ko-0.9.2
mediapipe==0.8.7.3
msgpack==1.0.3
mtcnn==0.1.1
numpy==1.23.4
oauthlib==3.2.2
opencv-contrib-python==4.5.3.56
opencv-python==4.8.1.78
openpyxl==3.0.10
opt-einsum==3.3.0
outcome==1.2.0
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
platformdirs==3.1.1
prometheus-client==0.16.0
prompt-toolkit==3.0.36
protobuf==3.20.0
psutil==5.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyfiglet==0.8.post1
pygame==2.1.2
pygraphviz==1.10
pylint==2.17.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyperclip==1.8.2
PyQt5==5.12.3
PyQt5-sip==12.9.0
pyqtgraph==0.10.0
pyshorteners==1.0.1
PySocks==1.7.1
python-crontab==2.7.1
python-dateutil==2.8.2
pytz==2022.4
pywatchman==1.4.1
PyYAML==6.0
redis==4.4.2
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
retina-face==0.0.13
rsa==4.9
scipy==1.7.1
screeninfo==0.8
selenium==4.7.2
service-identity==21.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.1
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.3.0
tokenizers==0.13.2
tomli==2.0.1
tomlkit==0.11.6
torch==1.13.1+cpu
torchaudio==0.13.1+cpu
torchvision==0.14.1+cpu
tornado==6.2
tqdm==4.64.1
transformers==4.26.1
trio==0.22.0
trio-websocket==0.9.2
Twisted==21.7.0
txaio==21.2.1
typing-extensions==4.0.0
tzdata==2022.7
urllib3==1.26.13
uWSGI==2.0.20
vine==5.0.0
wcwidth==0.2.6
werkzeug==3.0.0
wrapt==1.15.0
wsproto==1.2.0
xlrd==2.0.1
xlwt==1.3.0
you-get==0.4.1620
zipp==3.17.0
zope.interface==5.4.0
